Feature Extractor and Data Reader
2 feature extractors were added. The first one considers 32 features
which come from magnitudes and directions of the optical flow of the
video. the second one has 2 features which are the average of the
magnitudes and the average of the angles of the optical flow from the
video.

utils also contains a function capable of reading the path to all video
input files from the UCF action database.